Extract Cursor
